+++ lib/blocs/theme/theme_event.dart
@@ -1,0 +1,16 @@
+part of 'theme_bloc.dart';
+
+abstract class ThemeEvent extends Equatable {
+  const ThemeEvent();
+}
+
+class LoadThemeEvent extends ThemeEvent {
+  const LoadThemeEvent();
+  @override List<Object> get props => [];
+}
+
+class ChangeThemeEvent extends ThemeEvent {
+  final ThemeType theme;
+  const ChangeThemeEvent(this.theme);
+  @override List<Object> get props => [theme];
+}