+++ lib/data_providers/data/database_provider.dart
@@ -18,7 +18,7 @@ import 'package:flutter/material.dart';
 import 'package:path/path.dart' as p;
 import 'package:random_color/random_color.dart';
 import 'package:sqflite/sqflite.dart';
-import 'package:timecop/data_providers/data_provider.dart';
+import 'package:timecop/data_providers/data/data_provider.dart';
 import 'package:timecop/models/timer_entry.dart';
 import 'package:timecop/models/project.dart';