+++ lib/screens/dashboard/bloc/dashboard_bloc.dart
@@ -17,8 +17,8 @@ class DashboardBloc extends Bloc<DashboardEvent, DashboardState> {
 
   @override
   DashboardState get initialState {
-      Project newProject = projectsBloc.getProjectByID(settingsBloc.state.defaultProjectID);
-    return DashboardState("", newProject, false);
+    Project newProject = projectsBloc.getProjectByID(settingsBloc.state.defaultProjectID);
+    return DashboardState("", newProject, false, DateTime.now().subtract(Duration(days: 30)), null, <int>[], null);
   }
 
   @override
@@ -26,18 +26,35 @@ class DashboardBloc extends Bloc<DashboardEvent, DashboardState> {
     DashboardEvent event,
   ) async* {
     if(event is DescriptionChangedEvent) {
-      yield DashboardState(event.description, state.newProject, false);
+      yield DashboardState(event.description, state.newProject, false, state.filterStart, state.filterEnd, state.hiddenProjects, state.searchString);
     }
     else if(event is ProjectChangedEvent) {
-      yield DashboardState(state.newDescription, event.project, false);
+      yield DashboardState(state.newDescription, event.project, false, state.filterStart, state.filterEnd, state.hiddenProjects, state.searchString);
     }
     else if(event is TimerWasStartedEvent) {
       Project newProject = projectsBloc.getProjectByID(settingsBloc.state.defaultProjectID);
-      yield DashboardState("", newProject, true);
+      yield DashboardState("", newProject, true, state.filterStart, state.filterEnd, state.hiddenProjects, state.searchString);
     }
     else if(event is ResetEvent) {
       Project newProject = projectsBloc.getProjectByID(settingsBloc.state.defaultProjectID);
-      yield DashboardState("", newProject, false);
+      yield DashboardState("", newProject, false, state.filterStart, state.filterEnd, state.hiddenProjects, state.searchString);
+    }
+    else if(event is FilterStartChangedEvent) {
+      DateTime end = state.filterEnd;
+      if(state.filterEnd != null && event.filterStart.isAfter(state.filterEnd)) {
+        end = event.filterStart.add(Duration(hours: 23, minutes: 59, seconds: 59, milliseconds: 999));
+      }
+
+      yield DashboardState(state.newDescription, state.newProject, false, event.filterStart, end, state.hiddenProjects, state.searchString);
+    }
+    else if(event is FilterEndChangedEvent) {
+      yield DashboardState(state.newDescription, state.newProject, false, state.filterStart, event.filterEnd, state.hiddenProjects, state.searchString);
+    }
+    else if(event is FilterProjectsChangedEvent) {
+      yield DashboardState(state.newDescription, state.newProject, false, state.filterStart, state.filterEnd, event.projects, state.searchString);
+    }
+    else if(event is SearchChangedEvent) {
+      yield DashboardState(state.newDescription, state.newProject, false, state.filterStart, state.filterEnd, state.hiddenProjects, event.search);
     }
   }
 }