+++ lib/screens/dashboard/bloc/dashboard_event.dart
@@ -25,3 +25,27 @@ class TimerWasStartedEvent extends DashboardEvent {
   const TimerWasStartedEvent();
   @override List<Object> get props => [];
 }
+
+class FilterStartChangedEvent extends DashboardEvent {
+  final DateTime filterStart;
+  const FilterStartChangedEvent(this.filterStart);
+  @override List<Object> get props => [filterStart];
+}
+
+class FilterEndChangedEvent extends DashboardEvent {
+  final DateTime filterEnd;
+  const FilterEndChangedEvent(this.filterEnd);
+  @override List<Object> get props => [filterEnd];
+}
+
+class FilterProjectsChangedEvent extends DashboardEvent {
+  final List<int> projects;
+  const FilterProjectsChangedEvent(this.projects);
+  @override List<Object> get props => [projects];
+}
+
+class SearchChangedEvent extends DashboardEvent {
+  final String search;
+  const SearchChangedEvent(this.search);
+  @override List<Object> get props => [search];
+}