+++ lib/screens/dashboard/bloc/dashboard_state.dart
@@ -4,11 +4,47 @@ class DashboardState extends Equatable {
   final String newDescription;
   final Project newProject;
   final bool timerWasStarted;
-
-  const DashboardState(this.newDescription, this.newProject, this.timerWasStarted)
-    : assert(newDescription != null),
-      assert(timerWasStarted != null);
+  final DateTime filterStart;
+  final DateTime filterEnd;
+  final List<int> hiddenProjects;
+  final String searchString;
 
   @override
-  List<Object> get props => [newDescription, newProject, timerWasStarted];
+  bool get stringify => true;
+
+  const DashboardState(
+      this.newDescription,
+      this.newProject,
+      this.timerWasStarted,
+      this.filterStart,
+      this.filterEnd,
+      this.hiddenProjects,
+      this.searchString)
+      : assert(newDescription != null),
+        assert(timerWasStarted != null),
+        assert(hiddenProjects != null);
+
+  DashboardState.clone(
+    DashboardState state,
+    DateTime filterStart,
+    DateTime filterEnd,
+    String searchString,
+  {
+    String newDescription,
+    Project newProject,
+    bool timerWasStarted,
+    List<int> hiddenProjects,
+  })
+    : this(
+        newDescription ?? state.newDescription,
+        newProject ?? state.newProject,
+        timerWasStarted ?? state.timerWasStarted,
+        filterStart,
+        filterEnd,
+        hiddenProjects ?? state.hiddenProjects,
+        searchString
+      );
+
+  @override
+  List<Object> get props => [newDescription, newProject, timerWasStarted, filterStart, filterEnd, hiddenProjects, searchString];
 }