+++ test_driver/screenshots.dart
@@ -1,8 +1,8 @@
 import 'dart:async';
-import 'package:timecop/data_providers/data_provider.dart';
-import 'package:timecop/data_providers/mock_data_provider.dart';
-import 'package:timecop/data_providers/mock_settings_provider.dart';
-import 'package:timecop/data_providers/settings_provider.dart';
+import 'package:timecop/data_providers/data/data_provider.dart';
+import 'package:timecop/data_providers/data/mock_data_provider.dart';
+import 'package:timecop/data_providers/settings/mock_settings_provider.dart';
+import 'package:timecop/data_providers/settings/settings_provider.dart';
 import 'package:timecop/main.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_driver/driver_extension.dart';